added map to itinerary page with pins - work in progress
