+++ jsconfig.json
@@ -5,5 +5,5 @@
       "@/*": ["src/*"]
     }
   },
-  "include": ["next-env.d.js", "**/*.js", "**/*.jsx"]
+  "include": ["next-env.d.js", "**/*.js", "**/*.jsx", "src/components/autoCompleteElement"]
 }
